Create blackwidio riddle.py
it is preveious tcs code vita problem in this i learn  to form grid means matrix 2d to intilize the matrix as 0 s by row and column=" arr=[[0 for x in range(col)]for x in range(row)]"
and it take less time of execution and it also one of the problem in edyst daily challenge (april ). and update the array and to count common elements  according to the sum description .
Think nothing is difficult.